+++ src/components.d.ts
@@ -14,7 +14,6 @@ import { OptionValue as OptionValue1 } from "./components/zen-menu-item/zen-opti
 import { Offsets, Placement } from "@popperjs/core";
 import { StepItem } from "./components/zen-progress-tracker/zen-progress-tracker";
 import { StepsFilter } from "./components/zen-progress-tracker/types";
-import { TabItem, TabValue } from "./components/zen-tabs/zen-tabs";
 export namespace Components {
     interface ColorSwatch {
         /**
@@ -66,27 +65,31 @@ export namespace Components {
     }
     interface ZenAvatar {
         /**
-          * Show icon animation
+          * Set show icon animation
          */
         "animation": boolean;
         /**
-          * Users
+          * Set users input data
          */
         "users": AvatarData[];
     }
     interface ZenAvatarDetails {
         /**
-          * User email
+          * Set user email
          */
         "email": string;
         /**
-          * Icon background color
+          * Set icon background color
          */
         "iconBackground": string;
         /**
-          * Icon color
+          * Set icon color
          */
         "iconColor": string;
+        /**
+          * Set override for user name initials
+         */
+        "initials": string;
         /**
           * <Description generated in helper file>
          */
@@ -108,11 +111,11 @@ export namespace Components {
          */
         "paddingTop": Spacing;
         /**
-          * User name
+          * Set user name
          */
         "userName": string;
         /**
-          * Different variants
+          * Set fifferent variants
          */
         "variant": AvatarDetailVariant;
     }
@@ -136,27 +139,31 @@ export namespace Components {
     }
     interface ZenAvatarIcon {
         /**
-          * Background color
+          * Set background color
          */
         "background": string;
         /**
-          * Font color
+          * Set font color
          */
         "color": string;
         /**
-          * Email
+          * Set email
          */
         "email": string;
         /**
-          * Image URL
+          * Set image URL
          */
         "imageUrl": string;
         /**
-          * Icon size
+          * Set override for user name initials
+         */
+        "initials": string;
+        /**
+          * Set icon size
          */
         "size": AvatarIconSize;
         /**
-          * Name and Surname
+          * Set name and surname
          */
         "userName": string;
     }
@@ -794,33 +801,59 @@ export namespace Components {
     }
     interface ZenSpinner {
     }
+    interface ZenTab {
+        /**
+          * Set tab selected
+         */
+        "selected": boolean;
+    }
     interface ZenTable {
+        /**
+          * Space separated css grid columns<br/>(eg. `auto 1fr 1fr 200px 1fr`)
+         */
+        "columns": "";
     }
     interface ZenTableCell {
-    }
-    interface ZenTableHeader {
         /**
-          * Checkbox indeterminate state
+          * Is row expanded (read-only)
+         */
+        "$depth": number;
+        /**
+          * Can be expanded (read-only)
+         */
+        "$expandable": boolean;
+        /**
+          * Is row expanded (read-only)
+         */
+        "$expanded": boolean;
+        /**
+          * Cell is inside header (read-only)
+         */
+        "$header": boolean;
+        /**
+          * Checkbox indeterminate state (read-only)
          */
         "$indeterminate": boolean;
         /**
-          * Show checkbox
+          * Show checkbox (read-only)
          */
-        "selectable": boolean;
+        "$selectable": boolean;
         /**
-          * Select all rows
+          * Is row selected (read-only)
          */
-        "selected": boolean;
+        "$selected": boolean;
         /**
-          * Remains fixed at the top of the table during vertical scrolling
+          * Cell remains fixed at the top during scroll (mainly used for headers)
          */
-        "sticky": boolean;
-    }
-    interface ZenTableHeaderCell {
+        "$sticky": boolean;
         /**
-          * Remains fixed at the top of the table during vertical scrolling
+          * Cell custom background color
          */
-        "sticky": boolean;
+        "backgroundColor": string;
+        /**
+          * Spanns through whole table
+         */
+        "fullSpan": boolean;
     }
     interface ZenTableRow {
         /**
@@ -848,6 +881,10 @@ export namespace Components {
          */
         "hasRowsSelected": () => Promise<boolean>;
         /**
+          * Row represents header
+         */
+        "header": boolean;
+        /**
           * Returns elements parent row (depth -1) *
          */
         "parentRow": () => Promise<HTMLZenTableRowElement>;
@@ -860,6 +897,10 @@ export namespace Components {
          */
         "selected": boolean;
         /**
+          * Row remains fixed at the top during scroll (mainly used for headers)
+         */
+        "sticky": false;
+        /**
           * Visible if no depth or parent.expanded
          */
         "visible": boolean;
@@ -868,11 +909,7 @@ export namespace Components {
         /**
           * Index of currently selected tab.
          */
-        "tabs": TabItem[];
-        /**
-          * Index of currently selected tab.
-         */
-        "value": TabValue;
+        "value": number;
     }
     interface ZenText {
         /**
@@ -1254,6 +1291,12 @@ declare global {
         prototype: HTMLZenSpinnerElement;
         new (): HTMLZenSpinnerElement;
     };
+    interface HTMLZenTabElement extends Components.ZenTab, HTMLStencilElement {
+    }
+    var HTMLZenTabElement: {
+        prototype: HTMLZenTabElement;
+        new (): HTMLZenTabElement;
+    };
     interface HTMLZenTableElement extends Components.ZenTable, HTMLStencilElement {
     }
     var HTMLZenTableElement: {
@@ -1265,18 +1308,6 @@ declare global {
     var HTMLZenTableCellElement: {
         prototype: HTMLZenTableCellElement;
         new (): HTMLZenTableCellElement;
-    };
-    interface HTMLZenTableHeaderElement extends Components.ZenTableHeader, HTMLStencilElement {
-    }
-    var HTMLZenTableHeaderElement: {
-        prototype: HTMLZenTableHeaderElement;
-        new (): HTMLZenTableHeaderElement;
-    };
-    interface HTMLZenTableHeaderCellElement extends Components.ZenTableHeaderCell, HTMLStencilElement {
-    }
-    var HTMLZenTableHeaderCellElement: {
-        prototype: HTMLZenTableHeaderCellElement;
-        new (): HTMLZenTableHeaderCellElement;
     };
     interface HTMLZenTableRowElement extends Components.ZenTableRow, HTMLStencilElement {
     }
@@ -1348,10 +1379,9 @@ declare global {
         "zen-sortable-item": HTMLZenSortableItemElement;
         "zen-space": HTMLZenSpaceElement;
         "zen-spinner": HTMLZenSpinnerElement;
+        "zen-tab": HTMLZenTabElement;
         "zen-table": HTMLZenTableElement;
         "zen-table-cell": HTMLZenTableCellElement;
-        "zen-table-header": HTMLZenTableHeaderElement;
-        "zen-table-header-cell": HTMLZenTableHeaderCellElement;
         "zen-table-row": HTMLZenTableRowElement;
         "zen-tabs": HTMLZenTabsElement;
         "zen-text": HTMLZenTextElement;
@@ -1411,27 +1441,31 @@ declare namespace LocalJSX {
     }
     interface ZenAvatar {
         /**
-          * Show icon animation
+          * Set show icon animation
          */
         "animation"?: boolean;
         /**
-          * Users
+          * Set users input data
          */
         "users"?: AvatarData[];
     }
     interface ZenAvatarDetails {
         /**
-          * User email
+          * Set user email
          */
         "email"?: string;
         /**
-          * Icon background color
+          * Set icon background color
          */
         "iconBackground"?: string;
         /**
-          * Icon color
+          * Set icon color
          */
         "iconColor"?: string;
+        /**
+          * Set override for user name initials
+         */
+        "initials"?: string;
         /**
           * <Description generated in helper file>
          */
@@ -1453,11 +1487,11 @@ declare namespace LocalJSX {
          */
         "paddingTop"?: Spacing;
         /**
-          * User name
+          * Set user name
          */
         "userName"?: string;
         /**
-          * Different variants
+          * Set fifferent variants
          */
         "variant"?: AvatarDetailVariant;
     }
@@ -1481,27 +1515,31 @@ declare namespace LocalJSX {
     }
     interface ZenAvatarIcon {
         /**
-          * Background color
+          * Set background color
          */
         "background"?: string;
         /**
-          * Font color
+          * Set font color
          */
         "color"?: string;
         /**
-          * Email
+          * Set email
          */
         "email"?: string;
         /**
-          * Image URL
+          * Set image URL
          */
         "imageUrl"?: string;
         /**
-          * Icon size
+          * Set override for user name initials
+         */
+        "initials"?: string;
+        /**
+          * Set icon size
          */
         "size"?: AvatarIconSize;
         /**
-          * Name and Surname
+          * Set name and surname
          */
         "userName"?: string;
     }
@@ -2131,33 +2169,63 @@ declare namespace LocalJSX {
     }
     interface ZenSpinner {
     }
+    interface ZenTab {
+        /**
+          * Tab selected event
+         */
+        "onTabSelect"?: (event: CustomEvent<void>) => void;
+        /**
+          * Set tab selected
+         */
+        "selected"?: boolean;
+    }
     interface ZenTable {
+        /**
+          * Space separated css grid columns<br/>(eg. `auto 1fr 1fr 200px 1fr`)
+         */
+        "columns"?: "";
     }
     interface ZenTableCell {
-    }
-    interface ZenTableHeader {
         /**
-          * Checkbox indeterminate state
+          * Is row expanded (read-only)
+         */
+        "$depth"?: number;
+        /**
+          * Can be expanded (read-only)
+         */
+        "$expandable"?: boolean;
+        /**
+          * Is row expanded (read-only)
+         */
+        "$expanded"?: boolean;
+        /**
+          * Cell is inside header (read-only)
+         */
+        "$header"?: boolean;
+        /**
+          * Checkbox indeterminate state (read-only)
          */
         "$indeterminate"?: boolean;
         /**
-          * Show checkbox
+          * Show checkbox (read-only)
          */
-        "selectable"?: boolean;
+        "$selectable"?: boolean;
         /**
-          * Select all rows
+          * Is row selected (read-only)
          */
-        "selected"?: boolean;
+        "$selected"?: boolean;
         /**
-          * Remains fixed at the top of the table during vertical scrolling
+          * Cell remains fixed at the top during scroll (mainly used for headers)
          */
-        "sticky"?: boolean;
-    }
-    interface ZenTableHeaderCell {
+        "$sticky"?: boolean;
         /**
-          * Remains fixed at the top of the table during vertical scrolling
+          * Cell custom background color
          */
-        "sticky"?: boolean;
+        "backgroundColor"?: string;
+        /**
+          * Spanns through whole table
+         */
+        "fullSpan"?: boolean;
     }
     interface ZenTableRow {
         /**
@@ -2177,6 +2245,10 @@ declare namespace LocalJSX {
          */
         "expanded"?: boolean;
         /**
+          * Row represents header
+         */
+        "header"?: boolean;
+        /**
           * Row selected
          */
         "onRowSelectChanged"?: (event: CustomEvent<boolean>) => void;
@@ -2189,6 +2261,10 @@ declare namespace LocalJSX {
          */
         "selected"?: boolean;
         /**
+          * Row remains fixed at the top during scroll (mainly used for headers)
+         */
+        "sticky"?: false;
+        /**
           * Visible if no depth or parent.expanded
          */
         "visible"?: boolean;
@@ -2197,11 +2273,7 @@ declare namespace LocalJSX {
         /**
           * Index of currently selected tab.
          */
-        "tabs"?: TabItem[];
-        /**
-          * Index of currently selected tab.
-         */
-        "value"?: TabValue;
+        "value"?: number;
     }
     interface ZenText {
         /**
@@ -2413,10 +2485,9 @@ declare namespace LocalJSX {
         "zen-sortable-item": ZenSortableItem;
         "zen-space": ZenSpace;
         "zen-spinner": ZenSpinner;
+        "zen-tab": ZenTab;
         "zen-table": ZenTable;
         "zen-table-cell": ZenTableCell;
-        "zen-table-header": ZenTableHeader;
-        "zen-table-header-cell": ZenTableHeaderCell;
         "zen-table-row": ZenTableRow;
         "zen-tabs": ZenTabs;
         "zen-text": ZenText;
@@ -2462,10 +2533,9 @@ declare module "@stencil/core" {
             "zen-sortable-item": LocalJSX.ZenSortableItem & JSXBase.HTMLAttributes<HTMLZenSortableItemElement>;
             "zen-space": LocalJSX.ZenSpace & JSXBase.HTMLAttributes<HTMLZenSpaceElement>;
             "zen-spinner": LocalJSX.ZenSpinner & JSXBase.HTMLAttributes<HTMLZenSpinnerElement>;
+            "zen-tab": LocalJSX.ZenTab & JSXBase.HTMLAttributes<HTMLZenTabElement>;
             "zen-table": LocalJSX.ZenTable & JSXBase.HTMLAttributes<HTMLZenTableElement>;
             "zen-table-cell": LocalJSX.ZenTableCell & JSXBase.HTMLAttributes<HTMLZenTableCellElement>;
-            "zen-table-header": LocalJSX.ZenTableHeader & JSXBase.HTMLAttributes<HTMLZenTableHeaderElement>;
-            "zen-table-header-cell": LocalJSX.ZenTableHeaderCell & JSXBase.HTMLAttributes<HTMLZenTableHeaderCellElement>;
             "zen-table-row": LocalJSX.ZenTableRow & JSXBase.HTMLAttributes<HTMLZenTableRowElement>;
             "zen-tabs": LocalJSX.ZenTabs & JSXBase.HTMLAttributes<HTMLZenTabsElement>;
             "zen-text": LocalJSX.ZenText & JSXBase.HTMLAttributes<HTMLZenTextElement>;